recommender/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recommender"
version = "0.1.0"
description = "Travel recommendation service with Kafka consumer and gRPC server"
authors = [
    {name = "Team", email = "team@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "grpcio==1.50.0",
    "grpcio-tools==1.50.0",
    "kafka-python==2.0.2",
    "sqlalchemy==2.0.23",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "black==23.12.0",
    "ruff==0.1.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"
python_files = ["test_*.py", "*_test.py"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long handled by formatter

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py312"]
